# [SPARK-52737][CORE] Pushdown predicate and number of apps to FsHistoryProvider when listing applications