--- a/core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
+++ b/core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
@@ -212,18 +212,6 @@ class HistoryServer(
     provider.getListing()
   }
 
-  /**
-   * Returns a list of available applications, in descending order according to their end time.
-   *
-   * @param max The maximum number of applications to return
-   * @param predicate A function that filters the applications to be returned
-   * @return An iterator of matching applications up to the specified maximum
-   */
-  def getApplicationList(max: Int)(predicate: ApplicationInfo => Boolean)
-    : Iterator[ApplicationInfo] = {
-    provider.getListing(max)(predicate)
-  }
-
 
   def getEventLogsUnderProcess(): Int = {
     provider.getEventLogsUnderProcess()
@@ -237,9 +225,9 @@ class HistoryServer(
     getApplicationList()
   }
 
-  override def getApplicationInfoList(max: Int)(filter: ApplicationInfo => Boolean)
-    : Iterator[ApplicationInfo] = {
-    getApplicationList(max)(filter)
+  override def getApplicationInfoList(max: Int)(
+      filter: ApplicationInfo => Boolean): Iterator[ApplicationInfo] = {
+    provider.getListing(max)(filter)
   }
 
   def getApplicationInfo(appId: String): Option[ApplicationInfo] = {