--- a/core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
+++ b/core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
@@ -212,7 +212,6 @@ class HistoryServer(
     provider.getListing()
   }
 
-
   def getEventLogsUnderProcess(): Int = {
     provider.getEventLogsUnderProcess()
   }