@shardulm94 2025-07-16
Just as an update here, I am having trouble setting the scenario up locally at a scale where we can reproduce the issue. We originally identified the issue using flamegraphs from our production instance.